+++ app/api/auth/google/init/route.ts
@@ -3,6 +3,10 @@ import queryString from "query-string";
 import { NextRequest, NextResponse } from "next/server";
 
 export async function GET(req: NextRequest) {
+  const { searchParams } = new URL(req.url);
+  const nextPath = searchParams.get("nextPath");
+  console.log({ searchParams, nextPath });
+
   const { secret, csrfToken } = generateCSRFToken();
   console.log("googleClientId", process.env.GOOGLE_CLIENT_ID);
 
@@ -14,8 +18,6 @@ export async function GET(req: NextRequest) {
     redirect_uri: process.env.GOOGLE_REDIRECT_URI!,
     state: csrfToken,
   };
-
-  console.log({ options });
 
   const redirectUrl = `${googleOAuthURL}?${queryString.stringify(options)}`;
 
@@ -36,7 +38,8 @@ export async function GET(req: NextRequest) {
    * - to keep track of previous pages, limit to 2? 5? idkyetlol
    * - navigate user back to the previous page after login success
    */
-  const previousPage = req.headers.get("referer") || "/";
+  const previousPage = nextPath || req.headers.get("referer") || "/";
+
   response.cookies.set("previousPage", previousPage, {
     httpOnly: true,
     secure: process.env.NODE_ENV === "production",
